Laget modal på slette funksjonen.
