+++ src/components/editAndDelete/EditModal.tsx
@@ -1,7 +1,7 @@
 import { Modal } from "@navikt/ds-react";
 import React, { Dispatch, SetStateAction } from "react";
 import { EditTravel } from "./EditTravel";
-import { Travel } from "../models/Travel";
+import { Travel } from "../../models/Travel";
 
 export const EditModal = ({
   open,
